Fixed delivery tags to work with floats.

+++ delivery_database/delivery/templatetags/delivery_tags.py
@@ -5,6 +5,6 @@ register = template.Library()
 @register.simple_tag
 def divide(n, d):
   try:
-    return "{:.2f}".format((int(n)/int(d)) * 100)
+    return "{:.2f}".format((float(n)/float(d)) * 100)
   except (ValueError, ZeroDivisionError):
     return None
